Added PDF of the paper and the PPT
This paper wishes to highlight the impact that artificial intelligence has in the learning process of a student. Since we are now witnessing the pinnacle of an ever-expanding AI, we must learn to distinguish between what is useful and what brings more harm than good. The sheer amount of computational power and the ease in data processing must be used wisely in order for us to grow and evolve.